refactor(starpuff): 小怪 per-kind AI 機械式搬移 systems/enemyUpdates.ts
- 九種 per-kind 逐幀 AI（jelly/floaty/spiky/puffy/chompy/shelly/zappy/drilly/glowy）與
  行為常數整批搬移；enemies.ts 871 行降至 552，保留生成/傷害/回收/hazards 池與 spawnElite 入口
- 依賴面收斂 EnemyUpdateContext（target/elapsedMs 即時 getter＋pulseRing/spawnBite/popPuffy 回呼）
- 行為零改變：228 unit ＋混合彈/新怪 e2e characterization 保護通過

測試：typecheck、test:run 228 綠、e2e 抽驗 2 案綠
